+++ docs_snippets/lib/usages/widgets/item.dart
@@ -25,7 +25,7 @@ final item = FItem(
   onVariantChange: (previous, current) {},
   // {@endcategory}
   // {@category "Core"}
-  style: const .delta(margin: .value(.zero)),
+  style: const .delta(padding: .value(.zero)),
   enabled: true,
   selected: false,
   onPress: () {},
@@ -58,7 +58,7 @@ final raw = FItem.raw(
   onVariantChange: (previous, current) {},
   // {@endcategory}
   // {@category "Core"}
-  style: const .delta(margin: .value(.zero)),
+  style: const .delta(padding: .value(.zero)),
   enabled: true,
   selected: false,
   onPress: () {},